refactor(149152): reorganiza testes da regra ponto a ponto
