+++ src/pre_recebimento/__tests__/test_models.py
@@ -19,6 +19,8 @@ from ..documento_recebimento.models import (
 from ..ficha_tecnica.models import AnaliseFichaTecnica, FichaTecnicaDoProduto
 from ..layout_embalagem.models import LayoutDeEmbalagem, TipoDeEmbalagemDeLayout
 from ..qualidade.models import Laboratorio, TipoEmbalagemQld
+from src.pre_recebimento.cronograma_entrega.models import Cronograma
+from src.pre_recebimento.ficha_tecnica.models import FichaTecnicaDoProduto
 
 pytestmark = pytest.mark.django_db
 
@@ -417,3 +419,44 @@ def test_analise_ficha_tecnica_aprovada_categoria_flv(
         outras_informacoes_conferido=True,
     )
     assert analise.aprovada is True
+
+
+def test_ficha_tecnica_ponto_a_ponto_independente_da_categoria():
+    categorias = (
+        FichaTecnicaDoProduto.CATEGORIA_PERECIVEIS,
+        FichaTecnicaDoProduto.CATEGORIA_NAO_PERECIVEIS,
+        FichaTecnicaDoProduto.CATEGORIA_FLV,
+    )
+
+    for categoria in categorias:
+        ficha_tecnica = FichaTecnicaDoProduto(
+            categoria=categoria,
+            tipo_entrega=FichaTecnicaDoProduto.PONTO_A_PONTO,
+        )
+
+        assert ficha_tecnica.ponto_a_ponto is True
+
+
+def test_cronograma_identifica_ponto_a_ponto_pela_ficha_tecnica():
+    ficha_ponto_a_ponto = FichaTecnicaDoProduto(
+        categoria=FichaTecnicaDoProduto.CATEGORIA_PERECIVEIS,
+        tipo_entrega=FichaTecnicaDoProduto.PONTO_A_PONTO,
+    )
+    ficha_armazenavel = FichaTecnicaDoProduto(
+        categoria=FichaTecnicaDoProduto.CATEGORIA_FLV,
+        tipo_entrega=FichaTecnicaDoProduto.ARMAZEM,
+    )
+
+    cronograma_ponto_a_ponto = Cronograma(
+        ficha_tecnica=ficha_ponto_a_ponto,
+    )
+    cronograma_armazenavel = Cronograma(
+        ficha_tecnica=ficha_armazenavel,
+    )
+    cronograma_sem_ficha = Cronograma(
+        ficha_tecnica=None,
+    )
+
+    assert cronograma_ponto_a_ponto.ponto_a_ponto is True
+    assert cronograma_armazenavel.ponto_a_ponto is False
+    assert cronograma_sem_ficha.ponto_a_ponto is False
